+++ package.json
@@ -90,13 +90,14 @@
     "react-native-ble-plx": "^2.0.2",
     "react-native-config": "^1.4.1",
     "react-native-device-info": "^7.1.0",
-    "react-native-gesture-handler": "^1.8.0",
+    "react-native-gesture-handler": "^1.9.0",
+    "react-native-holy-sheet": "https://github.com/allenan/react-native-holy-sheet.git",
     "react-native-keyboard-aware-scroll-view": "^0.9.3",
     "react-native-localize": "^1.4.3",
     "react-native-navigation-bar-color": "^2.0.1",
     "react-native-onesignal": "^4.0.0-beta.4",
     "react-native-picker-select": "^8.0.3",
-    "react-native-reanimated": "^1.13.1",
+    "react-native-reanimated": "^2.0.0-alpha.9.2",
     "react-native-safe-area-context": "^3.1.8",
     "react-native-screens": "^2.12.0",
     "react-native-snap-carousel": "^4.0.0-beta.6",
